Roughly worked through the middle portion of GradTTS (the monotonic alignment) and began work on the diffusion U-Net decoder. Kind of stuck because of the required use of the GroupNormalization layer in the original work. I tried to copy/paste the raw implementation of the layer from a newer version of Tensorflow 2 but still found I had insufficient code/modules in the implementation. Will think about how to proceed.

+++ Grad-TTS-TF/model/utils.py
@@ -44,5 +44,5 @@ def generate_path(duration, mask):
 
 def duration_loss(logw, logw_, lengths):
 	loss = tf.math.reduce_sum((logw - logw_) ** 2) /\
-		tf.math.reduce_sum(lengths)
+		tf.math.reduce_sum(tf.cast(lengths, dtype=tf.float32))
 	return loss
